+++ pkg/shared/kvs/jetstream/kv_store_test.go
@@ -26,7 +26,8 @@ import (
 	"github.com/nats-io/nats.go"
 	"github.com/stretchr/testify/assert"
 
-	natstest "github.com/numaproj/numaflow/pkg/shared/clients/nats"
+	natsclass "github.com/numaproj/numaflow/pkg/shared/clients/nats"
+	natstest "github.com/numaproj/numaflow/pkg/shared/clients/nats/test"
 	"github.com/numaproj/numaflow/pkg/shared/kvs"
 )
 
@@ -39,7 +40,7 @@ func TestJetStreamKVStoreOperations(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -100,7 +101,7 @@ func TestJetStreamKVStoreWatch(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
 	defer cancel()
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -175,7 +176,7 @@ func TestJetStreamKVWithoutUpdates(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
 	defer cancel()
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -253,7 +254,7 @@ func TestJetStreamKVStoreErrorBinding(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	// Intentionally binding to a non-existing bucket to simulate error
@@ -270,7 +271,7 @@ func TestJetStreamKVStoreGetValueError(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -304,7 +305,7 @@ func TestJetStreamKVStoreListKeysError(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -338,7 +339,7 @@ func TestJetStreamKVStorePutKVError(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -372,7 +373,7 @@ func TestJetStreamKVStoreWatchError(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	js, err := testClient.JetStreamContext()
@@ -410,7 +411,7 @@ func TestJetStreamKVStoreWithContextDone(t *testing.T) {
 	s := natstest.RunJetStreamServer(t)
 	defer natstest.ShutdownJetStreamServer(t, s)
 
-	testClient := natstest.JetStreamClient(t, s)
+	testClient := natsclass.NewTestClientWithServer(t, s)
 	defer testClient.Close()
 
 	kvStore, err := NewKVJetStreamKVStore(ctx, kvName, testClient)
