+++ pkg/sideinputs/initializer/initializer_test.go
@@ -27,7 +27,8 @@ import (
 	"github.com/nats-io/nats.go"
 	"github.com/stretchr/testify/assert"
 
-	natstest "github.com/numaproj/numaflow/pkg/shared/clients/nats"
+	natsclass "github.com/numaproj/numaflow/pkg/shared/clients/nats"
+	natstest "github.com/numaproj/numaflow/pkg/shared/clients/nats/test"
 	"github.com/numaproj/numaflow/pkg/shared/kvs/jetstream"
 	"github.com/numaproj/numaflow/pkg/sideinputs/utils"
 )
@@ -58,7 +59,7 @@ func TestSideInputsInitializer_Success(t *testing.T) {
 	defer cancel()
 
 	// connect to NATS
-	nc := natstest.JetStreamClient(t, s)
+	nc := natsclass.NewTestClientWithServer(t, s)
 	defer nc.Close()
 
 	// create JetStream Context
@@ -127,7 +128,7 @@ func TestSideInputsTimeout(t *testing.T) {
 	defer cancel()
 
 	// connect to NATS
-	nc := natstest.JetStreamClient(t, s)
+	nc := natsclass.NewTestClientWithServer(t, s)
 	defer nc.Close()
 
 	// create JetStream Context
